+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## unreleased
+### Changes
+- Enabled automatic name conflict resolution by default
+  - https://github.com/StateSmith/StateSmith/issues/138
+
 ### Minor Fixes
 - js - remove incorrect type hinting for "enums"
   - https://github.com/StateSmith/StateSmith/issues/153